fix(detect): error on missing config.json / required topology fields (#22)
Surface the source of issue #22 (`could not broadcast array from shape:
[2560] to: [2048]` deep inside the extract pipeline) at the place the
config is loaded rather than at the matmul that finally trips on it.
Previously `read_config_json` returned `{}` when `config.json` was
absent, and `parse_model_config` fell back to magic defaults
(`hidden_size=2048`, `num_hidden_layers=32`, `intermediate_size=8192`).
For a real Gemma 3 4B model that produced a fake topology the extract
pipeline tried to broadcast against the real embed shape and panicked.

* `detect_architecture` now errors with `ModelError::ConfigMissing` when
  `config.json` is absent, and with `ModelError::ConfigFieldsMissing`
  when any of `hidden_size`, `num_hidden_layers`, `intermediate_size`
  is absent under both top level and `text_config`.
* `detect_from_json` stays infallible (keeps test ergonomics) but its
  topology fallbacks are 0 rather than guess values, and the validator
  catches the zero downstream.
* Remaining transformers-class defaults (num_attention_heads=8,
  num_key_value_heads=4, Gemma head_dim=256) are named constants
  instead of inline literals.
* `detect.rs` split into `detect/{mod,config_io,parser,tests}.rs` so
  each concern lives in its own file.

7 new unit tests cover: missing file, empty config, partial required
fields, nested `text_config`, flat config, mixed layout, validated-path
propagation. Existing 35 detect tests and the test_loading regression
all pass.
